python/todo-microservice-app/requirements.txt
cassandra-driver==3.27.0
Flask
Flask-Cors

